Update to use the latest scanner Docker image
When running the BOD scanning in a VPC, the concurrency I was using
before (550) seems to sometimes overwhelm the NAT gateway. In the
latest scanner Docker image the concurrency is reduced to 100.  We
still get good results and the scanning still completes in a
reasonable amount of time (about 6 hours).