Upgrade to the latest scanner, saver, and trustymail_reporter
The latest scanner uses the latest trustymail and domain-scan, which
add to the output the DMARC subdomain policy.

The latest saver includes the DMARC subdomain policy in the data that
is saved to the database.

The latest trustymail_reporter fails compliance for domains that have
a pct tag not equal to zero and/or have a subdomain policy not equal
to reject.