Use the latest trustymail_reporter Docker image
The latest version of trustymail_reporter now ignores case when
testing to see if a domain sends DMARC rua reports to the appropriate
email address for BOD 18-01 compliance.  This is the preferred
behavior, since modern email systems (including AWS SES, as I
verified) ignore case when looking at the local part of the email
address.