Modify to use the latest scanner Docker image
The latest scanner Docker image uses python:3.6-slim-buster as the
base image.  This fixes a bug where pshtt (and curl) were unable to
connect to any of the SSS domains.  This was because the old Ubuntu
base image was using an old version of gnutls.  The new base image
uses libopenssl instead of gnutls, and a much more recent vintage at
that.  See cisagov/scanner#34 for more details.